# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Noor Prayer Times API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PrayerTime {
  name: string;
  nameAr: string;
  /** Time in HH:MM format (24h) */
  time: string;
  enabled: boolean;
  /** Minutes after Adhan for Iqama (0 = no iqama displayed) */
  iqamaOffset?: number;
}

export interface NewsItem {
  id: string;
  text: string;
  textAr: string;
}

export interface HadithItem {
  hadith_ar: string;
  hadith_de: string;
}

/** @deprecated Use HadithItem */
export type AzkarItem = HadithItem;

export interface MosqueInfo {
  name: string;
  nameAr: string;
  address: string;
  addressAr: string;
  hijriAdjustment: number;
}

export interface PrayerData {
  mosque: MosqueInfo;
  prayers: PrayerTime[];
  news: NewsItem[];
  azkar: HadithItem[];
  lastUpdated: string;
}

export interface UpdatePrayerDataRequest {
  adminPassword: string;
  mosque: MosqueInfo;
  prayers: PrayerTime[];
  news: NewsItem[];
  azkar: HadithItem[];
}

export interface HadithUploadRequest {
  adminPassword: string;
  hadiths: HadithItem[];
}

export interface AdminVerifyRequest {
  password: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
}