# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Noor Prayer Times API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: prayer
    description: Prayer times and data
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /prayer-data:
    get:
      operationId: getPrayerData
      tags: [prayer]
      summary: Get prayer data
      description: Returns prayer times, news, and hadiths for the current configuration
      responses:
        "200":
          description: Prayer data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerData"

  /admin/prayer-data:
    put:
      operationId: updatePrayerData
      tags: [admin]
      summary: Update prayer data
      description: Update prayer times, news, and hadiths (requires admin password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrayerDataRequest"
      responses:
        "200":
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /prayer/hadith:
    post:
      operationId: uploadHadiths
      tags: [prayer]
      summary: Upload Hadith JSON
      description: Replace the hadith list with an uploaded JSON array (requires admin password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HadithUploadRequest"
      responses:
        "200":
          description: Hadiths saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/verify:
    post:
      operationId: verifyAdmin
      tags: [admin]
      summary: Verify admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyRequest"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    PrayerTime:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        time:
          type: string
          description: Time in HH:MM format (24h)
        enabled:
          type: boolean
        iqamaOffset:
          type: integer
          description: Minutes after Adhan for Iqama (0 = no iqama displayed)
      required:
        - name
        - nameAr
        - time
        - enabled

    NewsItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        textAr:
          type: string
      required:
        - id
        - text
        - textAr

    HadithItem:
      type: object
      description: A Prophetic Hadith with Arabic text and German translation
      properties:
        hadith_ar:
          type: string
          description: Arabic text of the hadith
        hadith_de:
          type: string
          description: German translation of the hadith
      required:
        - hadith_ar
        - hadith_de

    MosqueInfo:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        address:
          type: string
        addressAr:
          type: string
        hijriAdjustment:
          type: integer
      required:
        - name
        - nameAr
        - address
        - addressAr
        - hijriAdjustment

    PrayerData:
      type: object
      properties:
        mosque:
          $ref: "#/components/schemas/MosqueInfo"
        prayers:
          type: array
          items:
            $ref: "#/components/schemas/PrayerTime"
        news:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        azkar:
          type: array
          description: List of prophetic hadiths for the ticker
          items:
            $ref: "#/components/schemas/HadithItem"
        lastUpdated:
          type: string
      required:
        - mosque
        - prayers
        - news
        - azkar
        - lastUpdated

    UpdatePrayerDataRequest:
      type: object
      properties:
        adminPassword:
          type: string
        mosque:
          $ref: "#/components/schemas/MosqueInfo"
        prayers:
          type: array
          items:
            $ref: "#/components/schemas/PrayerTime"
        news:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        azkar:
          type: array
          items:
            $ref: "#/components/schemas/HadithItem"
      required:
        - adminPassword
        - mosque
        - prayers
        - news
        - azkar

    HadithUploadRequest:
      type: object
      properties:
        adminPassword:
          type: string
        hadiths:
          type: array
          items:
            $ref: "#/components/schemas/HadithItem"
      required:
        - adminPassword
        - hadiths

    AdminVerifyRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message